Improving README by explicitly stating unit of measure for limit
This took a while to track down and verify, had to look at the source,
then look at the Webpack docs which say "In the simple case, when only a
single loader is applied to the resource, the loader is called with one
parameter: the content of the resource file as string." That page
doesn't mention the `raw` option. Then discovered raw, discovered it was
a Buffer, and had to look up what Buffer.length is here:
https://nodejs.org/api/buffer.html#buffer_buf_length So I believe this
commit is correct.